requirements.txt
Bootstrap-Flask==1.5.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.20.1
pandas==1.2.1
protobuf==3.14.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
